src/requirements.txt
boto3==1.34.152
oracledb==2.2.1
python-dotenv==1.0.1
